requirements.txt
Flask
authlib
python-dotenv
mlflow